Call a remote REST API with the WebClient
In this step, we're using the reactive HTTP `WebClient` to call a remote
REST API. To do that, we first need to create an instance of that client
and configure it if filter functions (which in our case, will add HTTP
headers to the outgoing request).

We can then compose two REST calls into a single Publisher of
`ReactorIssue`s.

Note that because the templating engine we're using here does not
support `Publisher` types, we have to asynchronously resolve the whole
model before triggering the template rendering phase.